+++ src/babel/__init__.py
@@ -44,7 +44,7 @@ from babel.schema import (
     MetaParameters,
 )
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 
 
 def run(